refactor(api): give the posture, the run keys and a retry their own types
The record flattened `bsx_supervisor::{RootFs, Net, Display}` to strings the app compared against
literals, and a lease retried on a substring of a prose refusal. Now `control::Error::NotReady`,
record-side `Rootfs`/`Network`/`DisplayMode` and `RunId`/`RunName`; the on-disk format is unchanged.

+++ crates/app/src/frame.rs
@@ -71,7 +71,7 @@ impl Sinks {
 
     /// Records an upload of `frame_id` that took `took_ns` in `write_texture`: the log line is
     /// the frame id, the clock at the upload, and that duration.
-    fn record(&self, frame_id: u32, took_ns: u128) {
+    fn record(&self, frame_id: u32, took_ns: u64) {
         self.uploaded.fetch_add(1, Ordering::Relaxed);
         let mut log = self
             .log
